Create a PR when a Armada releases a new version
The only thing this GH action does currently is to create a new PR in the charts repository when a new version of Armada is cut via CircleCI.  

CircleCI will create a branch named `circleci-armada_*` and push it to this charts repository.  When this action sees any branch with a prefix `circleci`, it will run the `createArmadaPR` job and create a new PR for a human to review.