Add chart for siembol (#7)
Along with the quickstart.md in the siembol repository, this chart should help users spin up an installation of Siembol.

There are a few changes that need to be made after this merge.  Notably, we need to address the parameterization of `application.properties`.  The decision was made to merge this PR ahead of those changes.

Some of the changes since the original introduction of this PR:
* Adding basic job template for loading storm jars
* Use the latest siembol container image and its specific entrypoint
* Make imagePullPolicy configurable
* Make sure nimbus is available before trying to load the topology
* Add basic rules for alerting
* Remove gitlab and update storm and kafka charts
* Move config-editor templates from siembol GH repo to this chart
* add quickstart values
* remove topology job
* add /siembol/cache to zk setup job

Co-authored-by: Jonathan Giannuzzi <jonathan@giannuzzi.me>